fix display grimoire

+++ domain/magic/grimoire/grimoire.h
@@ -17,6 +17,6 @@ typedef struct Grimoire {
 Grimoire empty_grimoire();
 Grimoire start_grimoire();
 char* grimoire_to_string(Grimoire g);
-Spell get_spell_by_index(Grimoire g, uint8_t index);
+Spell get_spell_by_index(Grimoire g, int16_t index);
 
 #endif //DOOMDEPTHS_GRIMOIRE_H
